Add a size_t typedef.

+++ src/util/memcpy.h
@@ -2,8 +2,7 @@
 
 // Copy n bytes from src to dest. Return dest. The src and dest regions must not
 // overlap.
-static void* memcpy(void* restrict dest, const void* restrict src,
-                    unsigned int n) {
+static void* memcpy(void* restrict dest, const void* restrict src, size_t n) {
   char* o = dest;
   char* const end = o + n;
   const char* i = src;
