+++ src/util/strlen.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // Returns the length of a null-terminated string.
-static unsigned int strlen(const char* c_string) {
+static size_t strlen(const char* c_string) {
   const char* i = c_string;
   while (*i) i++;
   return i - c_string;
